Part4 - bloglist expansion - step9 - deleting a blog requires valid token

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
@@ -9,12 +8,10 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const token = request.token
-  const decodedToken = ( token ? jwt.verify(token, process.env.SECRET) : undefined )
-  if (!token || !decodedToken.id) {
+  if (!request.validToken) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(response.decodedToken.id)
 
   const blog = new Blog(request.body)
 
@@ -38,6 +35,9 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  if (!request.validToken) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = await Blog.findByIdAndDelete(request.params.id)
   if (blog) {
     response.status(204).end()
